WIP starting to score using .net
